requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
curl-cffi==0.6.2
feedparser==6.0.11
filelock==3.13.1
Flask==3.0.2
flask-cors==4.0.0
frozendict==2.4.0
google-generativeai==0.4.1
huggingface-hub==0.21.4
idna==3.6
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
tensorflow==2.16.1
urllib3==2.2.1
websockets==12.0
yfinance==0.2.37
